+++ app/src/main/java/de/thomaskuenneth/benice/AppChooserDialog.kt
@@ -30,7 +30,7 @@ fun AppChooserDialog(
                     columns = 1,
                     letterPosition = letterPosition,
                     showSpecials = true,
-                    onClick = onClick,
+                    onClick = { appInfo, _ -> onClick(appInfo)},
                     onLongClick = {},
                     selectBitmap = selectImage
                 )