+++ whatsnew.md
@@ -1,4 +1,5 @@
 # What's new
 
-- UI refinements
-- Control number of app icon columns on small, medium and large screens
+- Double tap to launch only this app
+- In app context menu, Launch adjacent launches the app adjacent to Be nice, whereas Launch launches only the selected app
+-